backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
better-profanity==0.7.0
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
gunicorn==20.1.0